Taking merged code for final testing

+++ quora-service/src/main/java/com/upgrad/quora/service/business/AdminService.java
@@ -28,21 +28,33 @@ public class AdminService {
      and the user that has to be deleted has valid accesstoken*/
 
     /*The annotation @Transactional esatblishs the connection, performs the operation and commits the transaction*/
-
     @Transactional(propagation = Propagation.REQUIRED)
     public String deleteUser(final String userUuid, final String authorization) throws AuthorizationFailedException, UserNotFoundException {
-
-
-                UserAuthTokenEntity userAuthTokenEntity = authorizationService.isValidActiveAuthTokenForAdmin(authorization);
+        /*Incorporated changes as per the changed User Dao method isValidAuthTokenForAdmin() */
+        UserAuthTokenEntity authTokenEntity = userDao.isValidActiveAuthTokenForAdmin(authorization);
+        /*checking if user has signed in*/
+        if (userDao.hasUserSignedIn(authorization)) {
+            /*Checking if the user has signed out and accordingly throwing an exception*/
+            if (authTokenEntity != null) {
+                /*Check if the logged in user has admin role or not*/
                 if (userDao.isRoleAdmin(authorization)) {
                     UserEntity userEntity = userDao.getUserById(userUuid);
-                    if(userEntity== null){
-                        throw new UserNotFoundException("USR-001","User with the entered Uuid to be deleted does not exist");
+                    /*Check if the user with the entered Uuid exists or not*/
+                    if (userEntity == null) {
+                        /*if user with the entered uuid does not exist throw an exception*/
+                        throw new UserNotFoundException("USR-001", "User with the entered Uuid to be deleted does not exist");
                     } else {
+                        /*if the user with the entered uuid exists, and the logged in user has admin role, we delete the given user*/
                         return adminDao.deleteUser(userUuid);
                     }
                 } else {
                     throw new AuthorizationFailedException("ATHR-003", "Unauthorized Access, Entered user is not an admin");
                 }
+            } else {
+                throw new AuthorizationFailedException("ATHR-002", "User is signed out");
             }
+        } else {
+            throw new AuthorizationFailedException("ATHR-001", "User has not signed in");
+        }
+    }
 }
